Split `fsagent.go` into separate Go files
This commit simplifies `fsagent.go` by moving a couple of parts to separate
files. It removes the global `done bool` and replaces it with a
`program`-specific channel.

For clarity, `goimports` was run to sort logically sort imports (first standard
library, second github resources, third "external" parties), and to
automatically set imports in the newly created files.
