+++ modules/mail.go
@@ -2,6 +2,7 @@ package modules
 
 import (
 	"crypto/tls"
+
 	"gopkg.in/gomail.v2"
 )
